requirements.txt
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorzero==2.0
docopt==0.6.2
gpiozero==2.0.1
idna==3.10
keyboard==0.13.5
mypy-extensions==1.0.0
numpy==2.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
requests==2.32.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
yarg==0.1.10
